Move commitment verification into verifier
